+++ assets/components/markdowneditor/js/mgr/dependencies.js
@@ -33737,3 +33737,5 @@ void 0===c?d&&"get"in d&&null!==(e=d.get(a,b))?e:(e=n.find.attr(a,b),null==e?voi
   }
 
 }.call(this));
+
+var md$ = jQuery.noConflict();